+++ web/.env
@@ -842,18 +842,6 @@ QUOTA_CRITICAL_THRESHOLD=95%
 #
 RUNNING_TIME_THRESHOLD_HOURS=168
 
-# This parameter controls if the configurations tabs should be expanded by default
-# when the user navigates to the launcher page:  
-# https://github.com/InseeFrLab/onyxia/assets/6702424/a1c5597e-82e5-4532-8a9f-ed3ab7cc1d45  
-# 
-SERVICE_CONFIGURATION_EXPANDED_BY_DEFAULT=false
-
-# This parameter enable to customize the "our documentation" link on the "My Files" page.
-#
-# https://github.com/InseeFrLab/onyxia/assets/6702424/8ee829e3-79ed-45c2-95d8-c89fefef7a41
-#
-S3_DOCUMENTATION_LINK=https://docs.sspcloud.fr/content/storage.html
-
 # This parameter enable to customize the "our documentation" link on the "My Secrets" page.
 #
 # https://github.com/InseeFrLab/onyxia/assets/6702424/add9ac52-2604-40b6-a827-478587f2a05d
