--- web/public/custom-resources-example/jsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "allowJs": true,
-        "checkJs": true,
-        "noUnusedLocals": false,
-        "noUnusedParameters": false
-    },
-    "include": ["./**/*.js", "../../src/vite-env.d.ts"]
-}
