--- web/public/custom-resources-example/ls3/docs/how-to-get-my-gitlab-token.md
@@ -1,3 +1,0 @@
-
-
-Link to account [here](/account/git).
