--- web/public/custom-resources-example/ls3/docs/new-user.md
@@ -1,6 +1,0 @@
-
-# Nouvel utilisateur 
-
-TODO.  
-
-[Tuto comment obtenir un token gitlab](%PUBLIC_URL%custom-resources-example/ls3/docs/how-to-get-my-gitlab-token.md)
