--- web/public/custom-resources/.gitignore
@@ -1,7 +1,0 @@
-# Ignore everything
-*
-
-# But not these files...
-!.gitignore
-!README.md
-!my-plugin.js
