--- web/public/custom-resources/my-plugin.js
@@ -1,8 +1,0 @@
-
-console.log("Initializing Onyxia plugin at /custom-resources/my-plugin.js");
-
-window.addEventListener("onyxiaready", ()=> {
-
-    console.log("Onyxia Global API ready", onyxia);
-
-});
