+++ web/src/vite-env.d.ts
@@ -57,8 +57,6 @@ type ImportMetaEnv = {
   QUOTA_WARNING_THRESHOLD: string
   QUOTA_CRITICAL_THRESHOLD: string
   RUNNING_TIME_THRESHOLD_HOURS: string
-  SERVICE_CONFIGURATION_EXPANDED_BY_DEFAULT: string
-  S3_DOCUMENTATION_LINK: string
   VAULT_DOCUMENTATION_LINK: string
   ONYXIA_API_URL: string
   DISABLE_DISPLAY_ALL_CATALOG: string
